requirements.txt
transformers==4.38.2
pandas>=2.2.1
joblib>=1.3.2
pytest>=8.0.2
torch==2.2.1
numpy==1.26.4
nltk==3.8.1
spacy==3.7.4
matplotlib==3.8.3
scikit-learn==1.4.1.post1
seaborn==0.13.2
lightning==2.2.1
jupyterlab==4.1.4
faiss-cpu==1.8.0
nlpaug==1.1.11
sacremoses==0.1.1
networkx==3.2.1